+++ src/kernel/uapi_errno.h
@@ -8,9 +8,11 @@
 
 #define EPERM          1
 #define ENOENT         2
+#define ESRCH          3
 #define EINTR          4
 #define EIO            5
 #define EAGAIN        11
+#define ECHILD        10
 #define EBADF          9
 #define ENOMEM        12
 #define EACCES        13